Do undo/redo at window level only

+++ NeoEdit.Editor/NESerialTracker.cs
@@ -2,7 +2,7 @@
 {
 	public static class NESerialTracker
 	{
-		public static int NESerial { get; private set; } = 1;
+		public static long NESerial { get; private set; } = 1;
 		public static void MoveNext() => ++NESerial;
 	}
 }
